Add SEO metadata for skill pages
